backend/requirements.txt
Flask==3.0.0
Flask-Cors ==4.0.0
loguru ==0.7.2
Flask-Compress== 1.14
psycopg2-binary== 2.9.9
Flask-RESTful ==0.3.10
pandas ==2.1.3
bcrypt==4.1.1
paho-mqtt ==1.6.1
PyJWT==2.8.0
pytz ==2023.3.post1
eventlet ==0.36.1
Flask-SocketIO ==5.3.6
requests ==2.31.0